+++ src/app.rs
@@ -3,22 +3,20 @@
 use std::borrow::Cow;
 use std::time::Instant;
 
-use crate::config::CONFIG;
-use crate::constants::ERROR_TIMEOUT;
-use crate::icon;
 use crate::message::Message;
 use crate::screenshot::RgbaHandle;
 use crate::selection::selection_lock::OptionalSelectionExt;
 use crate::theme::THEME;
+use iced::alignment::Vertical;
 use iced::keyboard::{Key, Modifiers};
-use iced::mouse::{Cursor, Interaction};
+use iced::mouse::Cursor;
 use iced::widget::canvas::Path;
-use iced::widget::{self, Action, canvas, stack};
-use iced::{Length, Point, Rectangle, Renderer, Size, Task, Theme, mouse};
+use iced::widget::text::Shaping;
+use iced::widget::{canvas, column, row, Space, Stack};
+use iced::{Background, Color, Element, Font, Length, Point, Rectangle, Size, Task};
 
 use crate::background_image::BackgroundImage;
 use crate::corners::{Side, SideOrCorner};
-use crate::mouse::MouseState;
 use crate::rectangle::RectangleExt;
 use crate::selection::{Selection, SelectionStatus};
 
@@ -58,8 +56,8 @@ use crate::selection::{Selection, SelectionStatus};
 pub static SAVED_IMAGE: std::sync::OnceLock<image::DynamicImage> = std::sync::OnceLock::new();
 
 /// Show an error message to the user
-#[derive(Debug, Clone, PartialEq, PartialOrd)]
-struct ErrorMessage {
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd)]
+pub struct ErrorMessage {
     /// Error message
     message: Cow<'static, str>,
     /// When the error was created
@@ -81,16 +79,16 @@ impl ErrorMessage {
 pub struct App {
     /// How many selections were created throughout the
     /// lifetime of the App
-    selections_created: usize,
+    pub selections_created: usize,
     /// The full screenshot of the monitor from which ferrishot was invoked
     /// We then create a window spanning the entire monitor, with this
     /// screenshot as background, with a canvas rendered on top - giving the
     /// illusion that we are drawing shapes on top of the screen.
-    screenshot: RgbaHandle,
+    pub screenshot: RgbaHandle,
     /// Area of the screen that is selected for capture
-    selection: Option<Selection>,
+    pub selection: Option<Selection>,
     /// Errors to display to the user
-    errors: Vec<ErrorMessage>,
+    pub errors: Vec<ErrorMessage>,
 }
 
 impl Default for App {
@@ -108,107 +106,92 @@ impl Default for App {
 }
 
 impl App {
-    /// Close the app
-    ///
-    /// This is like `iced::exit`, but it does not cause a segfault in special
-    /// circumstances <https://github.com/iced-rs/iced/issues/2625>
-    fn exit() -> Task<Message> {
-        iced::window::get_latest().then(|id| iced::window::close(id.expect("window to exist")))
-    }
-
-    /// Add a new error to the list of errors
-    fn error<T: Into<Cow<'static, str>> + std::fmt::Display>(&mut self, error: T) {
-        log::error!("Status Error: {error}");
-        self.errors.push(ErrorMessage::new(error));
-    }
-
-    /// Retrieve all valid errors
-    #[expect(dead_code, reason = "will be useful later")]
-    fn errors(&self) -> Vec<String> {
-        let now = Instant::now();
-        self.errors
-            .iter()
-            .rev()
-            .map_while(|err| {
-                let time_passed = now - err.timestamp;
-                (time_passed <= ERROR_TIMEOUT).then_some(err.message.to_string())
-            })
-            .collect()
-    }
-
-    /// Renders the black tint on regions that are not selected
-    fn render_shade(&self, frame: &mut canvas::Frame, bounds: Rectangle) {
-        let Some(selection) = self.selection.map(Selection::norm) else {
-            frame.fill_rectangle(bounds.position(), bounds.size(), THEME.non_selected_region);
-            return;
-        };
-
-        // represents the area outside of the selection
-        let outside = Path::new(|p| {
-            p.move_to(bounds.top_left());
-            p.line_to(bounds.top_right());
-            p.line_to(bounds.bottom_right());
-            p.line_to(bounds.bottom_left());
-            p.move_to(bounds.top_left());
-            p.move_to(selection.top_left());
-            p.line_to(selection.bottom_left());
-            p.line_to(selection.bottom_right());
-            p.line_to(selection.top_right());
-            p.move_to(selection.top_left());
-        });
-
-        frame.fill(&outside, THEME.non_selected_region);
-    }
-
-    /// Receives keybindings
-    #[must_use]
-    pub fn handle_key_press(key: Key, mods: Modifiers) -> Option<Message> {
-        match (key, mods) {
-            (Key::Named(iced::keyboard::key::Named::Escape), _) => Some(Message::Exit),
-            (Key::Character(ch), Modifiers::CTRL) if ch == "c" => Some(Message::CopyToClipboard),
-            (Key::Named(iced::keyboard::key::Named::Enter), _) => Some(Message::CopyToClipboard),
-            (Key::Character(ch), Modifiers::CTRL) if ch == "s" => Some(Message::SaveScreenshot),
-            (Key::Named(iced::keyboard::key::Named::F11), _) => Some(Message::FullSelection),
-            _ => None,
-        }
-    }
-
     /// Renders the app
     pub fn view(&self) -> iced::Element<Message> {
-        let icons = vec![
-            (
-                icon!(Fullscreen).on_press(Message::FullSelection).into(),
-                "Select entire monitor (F11)",
-            ),
-            (
-                icon!(Clipboard).on_press(Message::CopyToClipboard).into(),
-                "Copy to Clipboard (Enter)",
-            ),
-            (
-                icon!(Save).on_press(Message::SaveScreenshot).into(),
-                "Save Screenshot (Ctrl + S)",
-            ),
-            (icon!(Close).on_press(Message::Exit).into(), "Exit (Esc)"),
-        ];
+        Stack::new()
+            // taken screenshot in the background
+            .push(BackgroundImage::new(self.screenshot.clone().into()))
+            // border around the selection
+            .push(canvas(self).width(Length::Fill).height(Length::Fill))
+            // information popup, when there is no selection
+            .push_maybe(
+                self.selection
+                    .is_none()
+                    .then(|| self.render_welcome_message()),
+            )
+            // icons
+            .push_maybe(
+                // icons around the selection
+                self.selection.filter(|sel| sel.is_idle()).map(|sel| {
+                    let (image_width, image_height, _) = self.screenshot.raw();
+                    sel.render_icons(image_width as f32, image_height as f32)
+                }),
+            )
+            // size indicator
+            .push_maybe(self.selection.get().map(|(sel, key)| {
+                let (image_width, image_height, _) = self.screenshot.raw();
+                crate::widgets::size_indicator(image_height, image_width, sel.norm().rect, key)
+            }))
+            .into()
+    }
 
-        stack![
-            // the taken screenshot in the background
-            BackgroundImage::new(self.screenshot.clone().into()),
-            // the border around the selection
-            canvas(self).width(Length::Fill).height(Length::Fill),
-        ]
-        .push_maybe(
-            // icons around the selection
-            self.selection.filter(|sel| sel.is_idle()).map(|sel| {
-                let (width, height, _) = self.screenshot.raw();
-                sel.render_icons(icons, width as f32, height as f32)
-            }),
+    /// Renders the welcome message that the user sees when they first launch the program
+    fn render_welcome_message<'a>(&self) -> Element<'a, Message> {
+        use iced::widget::text;
+        const WIDTH: u32 = 340;
+        const HEIGHT: u32 = 100;
+
+        const FONT_SIZE: f32 = 13.0;
+
+        let (width, height, _) = self.screenshot.raw();
+        let vertical_space = Space::with_height(height / 2 - HEIGHT / 2);
+        let horizontal_space = Space::with_width(width / 2 - WIDTH / 2);
+
+        let bold = Font {
+            weight: iced::font::Weight::Bold,
+            ..Font::default()
+        };
+
+        let keys = |key: &'static str, action: &'static str| {
+            row![
+                row![
+                    Space::with_width(Length::Fill),
+                    text(key)
+                        .size(FONT_SIZE)
+                        .font(bold)
+                        .shaping(Shaping::Advanced)
+                        .align_y(Vertical::Bottom)
+                ]
+                .width(60.0),
+                Space::with_width(Length::Fixed(20.0)),
+                text(action).size(FONT_SIZE).align_y(Vertical::Bottom),
+            ]
+        };
+
+        let stuff = iced::widget::container(
+            column![
+                keys("Mouse", "Select screenshot area"),
+                keys("Ctrl + S", "Save screenshot to a file"),
+                keys("Enter", "Copy screenshot to clipboard"),
+                keys("Esc", "Exit"),
+            ]
+            .spacing(4.0)
+            .height(HEIGHT)
+            .width(WIDTH)
+            .padding(10.0),
         )
-        .push_maybe(self.selection.get().map(|(sel, key)| {
-            let (image_width, image_height, _) = self.screenshot.raw();
-            crate::widgets::size_indicator(image_height, image_width, sel.norm().rect, key)
-        }))
-        .into()
+        .style(|_| iced::widget::container::Style {
+            text_color: Some(THEME.fg_on_accent_bg),
+            background: Some(Background::Color(THEME.accent.scale_alpha(0.95))),
+            border: iced::Border::default(),
+            shadow: iced::Shadow {
+                color: Color::WHITE,
+                offset: iced::Vector::default(),
+                blur_radius: 6.0,
+            },
+        });
+
+        column![vertical_space, row![horizontal_space, stuff]].into()
     }
 
     /// Modifies the app's state
@@ -221,8 +204,10 @@ impl App {
             } => {
                 let sel = self.selection.unlock(sel_is_some);
                 let dy = new_height as f32 - sel.rect.height;
-                sel.rect.height = new_height as f32;
-                sel.rect.y -= dy;
+                *sel = sel
+                    .norm()
+                    .with_height(|_| new_height as f32)
+                    .with_y(|y| y - dy);
             }
             Message::ResizeHorizontally {
                 new_width,
@@ -230,8 +215,10 @@ impl App {
             } => {
                 let sel = self.selection.unlock(sel_is_some);
                 let dx = new_width as f32 - sel.rect.width;
-                sel.rect.width = new_width as f32;
-                sel.rect.x -= dx;
+                *sel = sel
+                    .norm()
+                    .with_width(|_| new_width as f32)
+                    .with_x(|x| x - dx);
             }
             Message::NoOp => (),
             Message::Exit => return Self::exit(),
@@ -244,7 +231,7 @@ impl App {
                             .map(|l| (cursor_pos, l, selected_region))
                     })
                 }) {
-                    let resized = SelectionStatus::Resized {
+                    let resized = SelectionStatus::Resize {
                         initial_rect: rect.norm().rect,
                         initial_cursor_pos: cursor,
                         resize_side: side,
@@ -252,7 +239,7 @@ impl App {
                     rect.status = resized;
                 } else if let Some((cursor, selected_region)) = self.cursor_in_selection_mut(cursor)
                 {
-                    let dragged = SelectionStatus::Dragged {
+                    let dragged = SelectionStatus::Move {
                         initial_rect_pos: selected_region.pos(),
                         initial_cursor_pos: cursor,
                     };
@@ -318,7 +305,6 @@ impl App {
 
                         return Self::exit();
                     }
-                    // TODO: show error to the user in a custom widget
                     Err(err) => {
                         self.error(format!("Could not copy the image: {err}"));
                     }
@@ -327,8 +313,7 @@ impl App {
             Message::SaveScreenshot => {
                 let Some(selection) = self.selection.as_ref().map(|sel| Selection::norm(*sel))
                 else {
-                    // TODO: instead of this, show an error to the user in
-                    // a custom widget
+                    self.error("Selection does not exist. There is nothing to copy!");
                     return ().into();
                 };
 
@@ -381,7 +366,7 @@ impl App {
                 let sel = self.selection.unlock(sel_is_some);
 
                 sel.rect = corners.resize_rect(selection.rect, cursor_pos, corner_point);
-                sel.status = SelectionStatus::Resized {
+                sel.status = SelectionStatus::Resize {
                     initial_rect: sel.rect,
                     initial_cursor_pos: cursor_pos,
                     resize_side: SideOrCorner::Corner(corners),
@@ -403,9 +388,46 @@ impl App {
         ().into()
     }
 
+    /// Renders the black tint on regions that are not selected
+    pub fn render_shade(&self, frame: &mut canvas::Frame, bounds: Rectangle) {
+        let Some(selection) = self.selection.map(Selection::norm) else {
+            frame.fill_rectangle(bounds.position(), bounds.size(), THEME.non_selected_region);
+            return;
+        };
+
+        // represents the area outside of the selection
+        let outside = Path::new(|p| {
+            p.move_to(bounds.top_left());
+            p.line_to(bounds.top_right());
+            p.line_to(bounds.bottom_right());
+            p.line_to(bounds.bottom_left());
+            p.move_to(bounds.top_left());
+            p.move_to(selection.top_left());
+            p.line_to(selection.bottom_left());
+            p.line_to(selection.bottom_right());
+            p.line_to(selection.top_right());
+            p.move_to(selection.top_left());
+        });
+
+        frame.fill(&outside, THEME.non_selected_region);
+    }
+
+    /// Receives keybindings
+    #[must_use]
+    pub fn handle_key_press(key: Key, mods: Modifiers) -> Option<Message> {
+        match (key, mods) {
+            (Key::Named(iced::keyboard::key::Named::Escape), _) => Some(Message::Exit),
+            (Key::Character(ch), Modifiers::CTRL) if ch == "c" => Some(Message::CopyToClipboard),
+            (Key::Named(iced::keyboard::key::Named::Enter), _) => Some(Message::CopyToClipboard),
+            (Key::Character(ch), Modifiers::CTRL) if ch == "s" => Some(Message::SaveScreenshot),
+            (Key::Named(iced::keyboard::key::Named::F11), _) => Some(Message::FullSelection),
+            _ => None,
+        }
+    }
+
     /// If the given cursor intersects the selected region, give the region and
     /// the cursor
-    fn cursor_in_selection(&self, cursor: Cursor) -> Option<(Point, Selection)> {
+    pub fn cursor_in_selection(&self, cursor: Cursor) -> Option<(Point, Selection)> {
         self.selection.and_then(|sel| {
             cursor
                 .position()
@@ -454,174 +476,35 @@ impl App {
                 .with_size(|_| Size { width, height })
         });
     }
-}
 
-impl canvas::Program<Message> for App {
-    type State = MouseState;
-
-    fn draw(
-        &self,
-        _state: &Self::State,
-        renderer: &Renderer,
-        _theme: &Theme,
-        bounds: Rectangle,
-        _cursor: mouse::Cursor,
-    ) -> Vec<canvas::Geometry> {
-        let mut frame = canvas::Frame::new(renderer, bounds.size());
-
-        self.render_shade(&mut frame, bounds);
-
-        if let Some(selection) = self.selection.map(Selection::norm) {
-            selection.render_border(&mut frame, THEME.accent);
-            selection.corners().render_circles(&mut frame, THEME.accent);
-        }
-
-        vec![frame.into_geometry()]
+    /// Close the app
+    ///
+    /// This is like `iced::exit`, but it does not cause a segfault in special
+    /// circumstances <https://github.com/iced-rs/iced/issues/2625>
+    fn exit() -> Task<Message> {
+        iced::window::get_latest().then(|id| iced::window::close(id.expect("window to exist")))
     }
 
-    fn mouse_interaction(
-        &self,
-        state: &Self::State,
-        _bounds: Rectangle,
-        cursor: iced::advanced::mouse::Cursor,
-    ) -> iced::advanced::mouse::Interaction {
-        self.selection
-            .and_then(|sel| {
-                // when we started dragging a side, even if we go outside of the bounds of that side (which
-                // happens often when we are dragging the mouse fast), we don't want the cursor to change
-                cursor
-                    .position()
-                    .and_then(|cursor| sel.corners().side_at(cursor).map(SideOrCorner::mouse_icon))
-                    // for example, if we start dragging top right corner, and move mouse to the
-                    // top left corner, we want the cursor to switch appropriately
-                    .or_else(|| {
-                        if let SelectionStatus::Resized { resize_side, .. } = sel.status {
-                            Some(resize_side.mouse_icon())
-                        } else {
-                            None
-                        }
-                    })
-            })
-            .unwrap_or_else(|| {
-                let is_left_released = state.is_left_released();
-                let is_moving_selection = self.selection.is_some_and(|sel| sel.status.is_dragged());
-
-                let is_grab = (is_left_released || is_moving_selection)
-                    && self.cursor_in_selection(cursor).is_some();
-                if is_grab {
-                    Interaction::Grab
-                } else {
-                    Interaction::Crosshair
-                }
-            })
+    /// Add a new error to the list of errors
+    fn error<T: Into<Cow<'static, str>> + std::fmt::Display>(&mut self, error: T) {
+        log::error!("Status Error: {error}");
+        self.errors.push(ErrorMessage::new(error));
     }
 
-    fn update(
-        &self,
-        state: &mut Self::State,
-        event: &iced::Event,
-        _bounds: Rectangle,
-        cursor: iced::advanced::mouse::Cursor,
-    ) -> Option<widget::Action<Message>> {
-        use iced::Event::Mouse;
+    /// Retrieve all valid errors
+    #[expect(dead_code, reason = "will be useful later")]
+    fn errors(&self) -> Vec<String> {
+        /// When there is an error, display it for this amount of time
+        const ERROR_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(5);
 
-        let message = match event {
-            Mouse(mouse::Event::ButtonPressed(mouse::Button::Left)) => {
-                state.left_click();
-                Message::LeftMouseDown(cursor)
-            }
-            Mouse(mouse::Event::ButtonPressed(mouse::Button::Right)) => {
-                state.right_click();
-                if let Some(cursor) = cursor.position() {
-                    if let Some((selection, sel_is_some)) = self.selection.get() {
-                        Message::ResizingToCursor {
-                            cursor_pos: cursor,
-                            selection: selection.norm(),
-                            sel_is_some,
-                        }
-                    } else {
-                        return None;
-                    }
-                } else {
-                    return None;
-                }
-            }
-            Mouse(mouse::Event::ButtonReleased(mouse::Button::Right)) => {
-                state.right_release();
-                Message::EnterIdle
-            }
-            Mouse(mouse::Event::ButtonReleased(mouse::Button::Left)) => {
-                state.left_release();
-                if CONFIG.instant && self.selections_created == 1 {
-                    Message::CopyToClipboard
-                } else {
-                    Message::EnterIdle
-                }
-            }
-            Mouse(mouse::Event::CursorMoved { position })
-                if self.selection.is_some_and(Selection::is_resized) =>
-            {
-                // FIXME: this will not be necessary when we have `let_chains`
-                let (selection, sel_is_some) =
-                    self.selection.get().expect("has `.is_some_and()` guard");
-
-                // FIXME: this will not be necessary when we have `let_chains`
-                let SelectionStatus::Resized {
-                    resize_side,
-                    initial_rect,
-                    initial_cursor_pos,
-                } = selection.status
-                else {
-                    unreachable!("has `.is_some_and(is_resized)` guard");
-                };
-
-                Message::InitialResize {
-                    current_cursor_pos: *position,
-                    resize_side,
-                    initial_cursor_pos,
-                    initial_rect,
-                    sel_is_some,
-                }
-            }
-            Mouse(mouse::Event::CursorMoved { position })
-                if state.is_left_clicked()
-                    && state.is_right_released()
-                    && self.selection.is_some_and(Selection::is_dragged) =>
-            {
-                // FIXME: this will not be necessary when we have `let_chains`
-                let current_selection = self.selection.expect("has `.is_some_and()` guard");
-
-                // FIXME: this will not be necessary when we have `let_chains`
-                let SelectionStatus::Dragged {
-                    initial_rect_pos,
-                    initial_cursor_pos,
-                } = current_selection.status
-                else {
-                    unreachable!();
-                };
-
-                Message::MovingSelection {
-                    current_cursor_pos: *position,
-                    initial_cursor_pos,
-                    current_selection,
-                    initial_rect_pos,
-                }
-            }
-            Mouse(mouse::Event::CursorMoved { position })
-                if state.is_left_clicked()
-                    && state.is_right_released()
-                    && self
-                        .selection
-                        .is_some_and(|sel| sel.is_idle() || sel.is_create()) =>
-            {
-                Message::ExtendNewSelection(*position)
-            }
-            Mouse(mouse::Event::ButtonPressed(mouse::Button::Middle)) => Message::FullSelection,
-            _ => return None,
-        };
-
-        log::info!("Received message: {message:#?}");
-
-        Some(Action::publish(message))
+        let now = Instant::now();
+        self.errors
+            .iter()
+            .rev()
+            .map_while(|err| {
+                let time_passed = now - err.timestamp;
+                (time_passed <= ERROR_TIMEOUT).then_some(err.message.to_string())
+            })
+            .collect()
     }
 }