+++ src/corners.rs
@@ -1,10 +1,7 @@
 //! `Corners` represents the 4 vertices of a `iced::Rectangle`
 use iced::{Point, Rectangle, mouse};
 
-use crate::{
-    constants::{FRAME_CIRCLE_RADIUS, FRAME_INTERACTION_AREA},
-    rectangle::RectangleExt as _,
-};
+use crate::rectangle::RectangleExt as _;
 
 /// Corner of a rectangle
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
@@ -17,6 +14,28 @@ pub enum Corner {
     BottomLeft,
     /// Bottom-right corner
     BottomRight,
+}
+
+/// Side of a rectangle
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum Side {
+    /// Top side
+    Top,
+    /// Right side
+    Right,
+    /// Bottom side
+    Bottom,
+    /// Left side
+    Left,
+}
+
+/// Side and corner
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum SideOrCorner {
+    /// One of the 4 sides of a rectangle
+    Side(Side),
+    /// One of the 4 corners of a rectangle
+    Corner(Corner),
 }
 
 impl Corner {
@@ -58,28 +77,6 @@ impl Corner {
             Self::BottomRight => initial_rect.with_width(|w| w + dx).with_height(|h| h + dy),
         }
     }
-}
-
-/// Side of a rectangle
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub enum Side {
-    /// Top side
-    Top,
-    /// Right side
-    Right,
-    /// Bottom side
-    Bottom,
-    /// Left side
-    Left,
-}
-
-/// Side and corner
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub enum SideOrCorner {
-    /// One of the 4 sides of a rectangle
-    Side(Side),
-    /// One of the 4 corners of a rectangle
-    Corner(Corner),
 }
 
 impl SideOrCorner {
@@ -137,6 +134,9 @@ impl Corners {
         frame: &mut iced::widget::canvas::Frame,
         accent_color: iced::Color,
     ) {
+        /// Radius of each of the 4 corner circles in the frame drawn around the selection
+        const FRAME_CIRCLE_RADIUS: f32 = 6.0;
+
         for circle in [
             self.top_left,
             self.top_right,
@@ -151,6 +151,9 @@ impl Corners {
 
     /// Return the interaction side for a point, if exists
     pub fn side_at(&self, point: Point) -> Option<SideOrCorner> {
+        /// Shadow to apply to elements
+        /// The area around each side of the frame which allows that side to be hovered over and resized
+        pub const FRAME_INTERACTION_AREA: f32 = 35.0;
         let top = Rectangle {
             x: self.top_left.x,
             y: self.top_left.y - FRAME_INTERACTION_AREA / 2.,